requirement.txt
altgraph==0.10.2
amqp==2.2.2
anyjson==0.3.3
apipkg==1.4
appnope==0.1.0
asana==0.6.2
astroid==1.4.9
awsebcli==3.5.3
backports-abc==0.4
backports.functools-lru-cache==1.3
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bdist-mpkg==0.5.0
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
billiard==3.5.0.3
bleach==1.4.3
blessed==1.9.5
#bonjour-py==0.3
botocore==1.2.3
browsermob-proxy==0.8.0
Cartridge==0.11.0
celery==4.0.0
cement==2.4.0
certifi==2016.2.28
cffi==1.8.3
chardet==2.3.0
colorama==0.3.3
contextdecorator==0.10.0
contextlib2==0.5.4
cryptography==1.5.2
decorator==4.0.10
dj-database-url==0.3.0
Django==1.9.4
django-celery-beat==1.0.1
django-contrib-comments==1.7.1
django-dbbackup==3.1.3
django-debug-toolbar==1.8
django-extensions==1.6.7
django-filter==0.13.0
django-json-response==1.1.5
django-markdown-deux==1.0.5
django-master-password==1.1.1
django-multiselectfield==0.1.5
django-pdb==0.5.1
django-postgrespool==0.3.0
django-redis==4.8.0
django-redis-cache==1.7.1
django-subdomains==2.1.0
django-tagging==0.4.5
django-webpack-loader==0.3.0
djangorestframework==3.3.3
docker==2.2.1
docker-py==1.1.0
docker-pycreds==0.2.1
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
dropbox==7.1.1
EasyProcess==0.2.2
email==4.0.2
entrypoints==0.2.2
enum34==1.1.6
eventlet==0.18.4
execnet==1.4.1
fake-factory==0.5.9
filebrowser-safe==0.4.3
flake8==3.0.2
functools32==3.2.3.post2
future==0.15.2
get==0.0.11
gitdb==0.6.4
GitPython==1.0.2
grappelli-safe==0.4.2
greenlet==0.4.10
gspread==0.4.1
gunicorn==19.4.5
hiredis==0.2.0
html5lib==0.9999999
httplib2==0.9.2
idna==2.1
ImageScraper==2.0.7
instamojo-wrapper==1.1
ipaddress==1.0.16
ipdb==0.10.2
ipykernel==4.3.1
ipython==5.0.0
ipython-genutils==0.1.0
ipywidgets==5.1.5
isort==4.2.5
Jinja2==2.8
jmespath==0.7.1
jsonfield==1.0.3
jsonschema==2.5.1
jupyter-client==4.3.0
jupyter-core==4.1.0
kombu==4.1.0
lazy-object-proxy==1.2.2
#LBForum==0.9.23
lxml==3.6.0
macholib==1.5.1
Markdown==2.6.6
markdown2==2.3.1
MarkupSafe==0.23
#matplotlib==1.3.1
mccabe==0.5.0
Mezzanine==4.1.0
mistune==0.7.3
modulegraph==0.10.4
namedlist==1.7
nbconvert==4.2.0
nbformat==4.0.1
newrelic==2.90.0.75
nose-parameterized==0.5.0
notebook==4.2.1
#numpy==1.8.0rc1
oauth2client==4.0.0
oauthlib==1.1.2
olefile==0.44
orderedset==2.0
pathlib2==2.1.0
pathspec==0.3.3
pexpect==4.2.0
pickleshare==0.7.2
Pillow==3.2.0
post==0.0.8
prompt-toolkit==1.0.3
#psycopg2==2.6
ptyprocess==0.5.1
public==0.0.34
py==1.4.31
py2app==0.7.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycodestyle==2.0.0
pycparser==2.16
pycrypto==2.6.1
pyflakes==1.2.3
PyGithub==1.26.0
Pygments==2.1.3
PyJWT==1.4.1
pylint==1.6.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==16.2.0
pyparsing==2.0.1
PyPDF2==1.26.0
pytest==2.9.0
pytest-allure-adaptor==1.7.2
pytest-html==1.8.0
pytest-rerunfailures==2.0.0
pytest-xdist==1.15.0
python-dateutil==2.4.2
python-http-client==2.2.1
python-openid==2.2.5
python-resize-image==1.1.10
python-social-auth==0.2.19
pytz==2016.4
PyVirtualDisplay==0.2
PyYAML==3.11
pyzmq==15.3.0
query-string==0.0.8
raven==5.32.0
redis==2.10.5
redisearch==0.6.3
reportlab==3.3.0
request==0.0.8
requests==2.10.0
requests-oauthlib==0.6.1
rmtest==0.5.2
rsa==3.4.2
#scipy==0.13.0b1
selenium==2.53.6
sendgrid==3.6.3
setproctitle==1.1.10
simplegeneric==0.8.1
SimplePool==0.1
singledispatch==3.4.0.3
six==1.10.0
slackclient==1.0.5
smmap==0.9.0
SQLAlchemy==1.0.4
sqlparse==0.2.3
terminado==0.6
texttable==0.8.3
tinify==1.3.0
tornado==4.3
traitlets==4.2.2
typing==3.5.3.0
tzlocal==1.2.2
UnittestZero==0.2.0
urllib3==1.19.1
vine==1.1.4
virtualenv==13.1.2
wcwidth==0.1.5
websocket-client==0.32.0
whitenoise==1.0.6
widgetsnbextension==1.2.3
wrapt==1.10.10
xattr==0.6.4
xhtml2pdf==0.0.6
yapf==0.16.3
youtube-dl==2016.2.22
zope.interface==4.1.1
